+++ src/Factory/FilesFactory.php
@@ -3,7 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Factory;
 
-use Innmind\Http\Message\Files;
+use Innmind\Http\ServerRequest\Files;
 
 /**
  * @psalm-immutable