+++ src/Factory/Query/QueryFactory.php
@@ -5,8 +5,7 @@ namespace Innmind\Http\Factory\Query;
 
 use Innmind\Http\{
     Factory\QueryFactory as QueryFactoryInterface,
-    Message\Query,
-    Message\Query\Parameter,
+    ServerRequest\Query,
 };
 
 /**