+++ src/Request/Stringable.php
@@ -1,0 +1,52 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\Http\Request;
+
+use Innmind\Http\Request;
+use Innmind\Filesystem\File\Content;
+use Innmind\Immutable\{
+    Sequence,
+    Str,
+};
+
+/**
+ * @psalm-immutable
+ */
+final class Stringable
+{
+    private function __construct()
+    {
+    }
+
+    public function __invoke(Request $request): Content
+    {
+        $status = Str::of("%s %s HTTP/%s\n")->sprintf(
+            $request->method()->toString(),
+            $request->url()->toString(),
+            $request->protocolVersion()->toString(),
+        );
+        $headers = $request
+            ->headers()
+            ->all()
+            ->map(static fn($header) => $header->toString())
+            ->map(Str::of(...))
+            ->map(static fn($header) => $header->append("\n"));
+        $body = $request->body()->chunks();
+
+        return Content::ofChunks(
+            Sequence::lazyStartingWith($status)
+                ->append($headers)
+                ->add(Str::of("\n"))
+                ->append($body),
+        );
+    }
+
+    /**
+     * @psalm-pure
+     */
+    public static function new(): self
+    {
+        return new self;
+    }
+}