+++ src/Response/StatusCode/Range.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\Http\Response\StatusCode;
+
+enum Range
+{
+    case informational;
+    case successful;
+    case redirection;
+    case clientError;
+    case serverError;
+}