+++ src/Sender.php
@@ -3,8 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Http;
 
-use Innmind\Http\Message\Response;
-
 interface Sender
 {
     public function __invoke(Response $response): void;